requirements.txt
python==3.10
tensorflow>=2.10.0
numpy>=1.21.0
Pillow>=9.0.0
scipy>=1.7.0
matplotlib>=3.4.0
tensorflow-addons>=0.17.0
flask>=2.0.0
werkzeug>=2.0.0
flask-login>=0.6.2
flask-pymongo>=2.3.0
pymongo>=4.3.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
dnspython>=2.3.0
email-validator>=2.0.0
itsdangerous==2.1.
Flask-WTF==1.0.0